LeapYear.java
Given a year, find if it is a leap year.
A Leap year is the year that is multiple of 4. However, multiples of 100 except for the multiples of 400 are not leap years.
Your task is to complete the function isLeapYear which receives the year as its parameter and returns 1 or 0 depending upon if the year is a leap year or not.