Find power of a number
Ninja is sitting in an examination hall. He is encountered with a problem statement, "Find ‘X’ to the power ‘N’ (i.e. ‘X’ ^ ‘N’). Where ‘X’ and ‘N’ are two integers." Ninja was not prepared for this question at all, as this question was unexpected in the exam. He is asking for your help to solve this problem. Help Ninja to find the answer to the problem.